Reorganize password page helper files
